Rewrite diff's line comparison and simple inline diff
`compare.rs` and `inline.rs`, together because the first's shape decides
the second's: every line the diff compares arrives as a **NUL-terminated**
buffer, so the byte range of one is a slice whose every tail still has a
valid `as_ptr()`. That is what lets the three mbyte calls keep their
pointer signatures while all the arithmetic around them becomes ordinary
indexing.

`diff_cmp` is now `lines_equal(&CStr, &CStr) -> bool`. Nothing ever read
its `strcmp` sign, and taking references moves the NUL guarantee into the
type: `strcmp(s1, s2) == 0` is `s1.to_bytes() == s2.to_bytes()`, and both
call sites already had the reference to hand. `diff_equal_char`'s
out-parameter becomes `char_len(&[u8], &[u8]) -> Option<usize>`.

Two spellings of the C are load-bearing and are documented rather than
tidied: an *empty* tail has `utfc_ptr2len` 0, which falls into the
single-byte arm and compares NUL against NUL, so two lines that have both
ended report a shared character of length 1 -- `common_suffix`'s walk
depends on that final joint step back. And `icase` folds a multibyte
character with `utf_fold` but a single-byte one with the locale's
`tolower`, which are not the same map.

`diff_find_change_simple`'s 139 lines are three items: `common_prefix`
and `common_suffix`, both **safe** functions over slices, and a 30-line
body that owns the copy and the accumulation. The two walks advance their
sides independently -- under `iwhite` a run of white space matches a run
of a different length -- which is much easier to see with two indices
than with two pointers into different lines.

`diff_change_parse` and `diff_find_change`'s "is this line an addition"
tests are both `(0..DB_COUNT).all(..)`; `f_diff_hlID`'s three-way
`linestatus` test is a `match` whose arms name what each sign means.
Two `xstrdup`/`xfree` pairs become `CStr::to_owned()`, with the comment
the C carries about *why* the copy exists (the next `ml_get_buf`
invalidates the buffer) kept at both sites.

`--allow-growth`, one row: `compare.rs` 3 -> 4. `head_off` is a safe
wrapper over `utf_head_off` shared by `common_prefix` and
`common_suffix`, which is the ratchet's known "a helper shared by two
functions has to pay one unit" shape -- the alternative spellings all
cost the same or more (a closure per site is two blocks; an `unsafe fn`
per walk is two). It buys ~100 lines of pointer arithmetic becoming slice
code. Family total 77 -> 78.

Gates: diffsweep byte-identical; clippy 2,929 -> 2,929; `cargo-test` 676
(+2, `skip_white` and `byte_at`).

+++ crates/nvim/src/nvim/diff/fold.rs
@@ -168,7 +168,7 @@ pub unsafe extern "C" fn diff_check_with_linestatus(
                         && !(*curtab.get()).tp_diffbuf[i_0 as usize].is_null()
                         && (*dp).df_count[i_0 as usize] != 0 as linenr_T
                     {
-                        if !diff_equal_entry(dp, idx, i_0) {
+                        if !diff_equal_entry(dp, idx as usize, i_0 as usize) {
                             if !linestatus.is_null() {
                                 *linestatus = -1 as ::core::ffi::c_int;
                             }
